backend/requirements.txt
# Backend Requirements - Essential packages only

Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.9.0
